+++ CHANGES.md
@@ -3,6 +3,7 @@ GH::watch changelog
 
 ## v1.17 (ToDo)
 - #74 - bugfix of authentication fail on new device when old token exists on github.com  
+- #68 - bugfix for rare exception when notification list screen is rotated during loading info to show notification  
 
 ## v1.16 (Jan 28th, 2015)
 - #67 - bugfix of NPE during widget add and remove